% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hac_table_e}
\alias{hac_table_e}
\title{HAC 16 neonatal birth trauma numerator codes}
\format{
Same column structure as \code{hac_table_a}.
}
\source{
ACSQHC HAC specification Excel files.
}
\usage{
hac_table_e
}
\description{
ICD-10-AM codes for the HAC 16 numerator (neonatal birth trauma). One row
per code x ICD edition.
}
\keyword{datasets}
