% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10am.hacs-package.R
\docType{package}
\name{icd10am.hacs-package}
\alias{icd10am.hacs}
\alias{icd10am.hacs-package}
\title{icd10am.hacs: Hospital Acquired Complications for ICD-10-AM Coded Data}
\description{
Identifies Hospital Acquired Complications (HACs) in Australian admitted patient care data coded using ICD-10-AM, using the ACSQHC HAC classification.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/healthpolicy/icd10am.hacs}
  \item Report bugs at \url{https://github.com/healthpolicy/icd10am.hacs/issues}
}

}
\author{
\strong{Maintainer}: Jim Pearse \email{ejpearse@gmail.com}

}
\keyword{internal}
